requirements.txt
openai>=1.45.0
httpx>=0.27.2
certifi>=2024.8.30
python-dotenv>=1.0.1
tiktoken>=0.7.0
python-docx>=1.1.2
pdfplumber>=0.11.4
